CSTX Refactor: fix tests
