+++ pkg/cstx/handler.go
@@ -8,5 +8,6 @@ import (
 type handler interface {
 	PublishCSXTToQueue(task structs.RMQPublishRequestTask, cstx CrossServiceTransaction) errs.APIError
 	PublishCSXTToExchange(task structs.PublishToExchangeTask, cstx CrossServiceTransaction) errs.APIError
+
 	PublishToExchange(task structs.PublishToExchangeTask, additionalHeaders ...structs.RMQHeader) errs.APIError
 }
