+++ tests/errors.go
@@ -1,12 +1,8 @@
-package rmqworker
+package tests
 
 import (
-	"errors"
-
 	"github.com/matrixbotio/rmqworker-lib/pkg/errs"
 )
-
-var ErrCSTXTimeout = errors.New("crossServiceTransaction timeout")
 
 func ConvertAPIError(err errs.APIError) error {
 	if err != nil {
